refactor(functions): separate twitter api functions
