+++ packages/functions/src/utils/twitter/index.ts
@@ -1,0 +1,6 @@
+export interface TwitterUserInterface {
+  id_str: string;
+  screen_name: string;
+  name: string;
+  profile_image_url_https: string;
+}
